requirements.txt
Flask==3.0.3
python-telegram-bot==20.0
python-dotenv==1.0.1
gunicorn
